Updated Movement to work with TileMapColliders instead of just BoxColliders.
Movement still not where I would like it, but is working. 4 directional movement works along with 4 directional gravity. Idealy, I would like 16 directional gravity (at a minimum). Also, when moving left or right into a wall then going over it without releasing left or right, velocity is not starting from 0%, but instead is at 100%. This needs to be fixed.