Changed indicators, fixed minor physics bug.
Indicators now reguire a colour and amount. For barriers that require more than 1 button of a certain colour, have that many indicators with each requiring an amount 1 larger than the previous. Eg for a door that needs 3 green buttons, we have 3 green indicators, one with an amount of 1, the other with an amount of 2, and the last with an amount of 3.